push everything related to thread to sidebar
